Fix AutoComplete from showing up when typing Emails (#1574)
* Fix AutoComplete from showing up when typing Emails

The AutoComplete View pops up when typing '@' to mention usernames, but when typing email addresses the same View pops up. This was fixed by checking whether the previous letter before the '@' is an empty space. This stops the view from showing as the letter before '@' would never be a space in an email.

Fixes:#1553

* Mention can follow a blank space, hash and a colon

If we enforce that the @ should be preceded by a blank space, then ':@abc' would look for an emoji with @abc. To prevent that we add ':' and '#' with the space.

* Fix Failing Tests