flow: Use `import type` in a bunch of places we should.
In each of these places (all importing from `types`, we're importing a
type for Flow to see -- not a real object for use at runtime.  So the
way we're supposed to spell that is `import type`.

We've been getting away with this so far because `types` is... not
actually typed!  So as far as Flow as concerned, it might actually
contain real objects named `User`, `Message`, etc.  We're about to
make `types` typed in order to get better type-checking across our
codebase; so this is one of the things we need to fix to do that.