vscode: Configure to use Flow and not TypeScript.
I'm not sure if we ultimately want to prefer VS Code or Atom/Nuclide,
but for now I want to try to make them both work as well as possible.

With these settings, VS Code stops giving errors for seeing Flow
syntax in .js files; and on the other hand, starts making use of Flow
to provide type information, parameter hints, context-aware
navigation, and so on.