fix: Replace data in narrow on anchor 0 or Max value
We used 'replaceExisting` value to determine if we should replace
current messages in fetched narrow with new messages.

This missed several key cases; we used it only on startup.

All cases where we need to replace data have an anchor of 0 or
Number.MAX_SAFE_INTEGER.

This is a fix for several erratic message list behaviors.