navReducers: Fix types to acknowledge state could be null.
The router's `getStateForAction` method returns null if it finds
nothing.  Given that we sometimes call it with an arbitrary action, an
assert that that doesn't happen doesn't seem appropriate; instead,
acknowledge that that can happen by using a "maybe type", with `?`.

This fact means any code relying on this state needs to make sure to
check before doing something that would break on null.  Marking the
type with `?` should mean Flow finds those places for us by giving an
error, and the spot fixed up in this commit is the only one it found.
Surely this can't be the only place we use this state... so hopefully
we're already checking elsewhere.  If not, this is just another case
where better Flow type coverage would make our code more robust.