flow: Enable Flow and make it pass in 'js.js'
These are the minimal changes required to make Flow tests pass.

document.body can be 'null' by specification so Flow is strict with
its checks. While it is extremely unlikely that this happens,
we add some checks to make Flow happy.

The definitions for addEventListener are imperfect and Flow complains.
Making them correct is too complicated so we just ignore the issues.

Also, we add a check for #typing element existance.
This, in addition to making Flow happy, is a good idea.