Make progressiveTimeout test less flaky
This test deals with async timing, for which I had some reservations.
It seems I was right. This seems to fail from time to time when run
on CI. I did relax the checks and removed one of the upper limit ones
(< 2050).

If this version fails even once, we should remove the test.