docs: Expand and correct docs on making an iOS release.
Add a description of exactly how to get the release out to our
TestFlight users.  This is important and we've missed it for some
releases in the past, so hopefully this will make it easier.

Also, a couple of steps I encountered in my first upload and
hoped were first-upload-only turn out to apply every time.