webview: Send a 'ready' message from webview js
We use postMessage to send messages to and from the webview.
Custom versions of that standard functions are injected by the
WebView component.

Sometimes, on Android (and seemingly not on iOS) it takes some time
for this custom JS to get injected, and we send our initial message
to the webview before it is ready. Thus the first message is not
delivered and often results in the 'infamous loading messages bug'

By checking for the function parameter count (window.postMessage.length === 1)
we can differentiate if the custom function is injected already or not.
Then we send a custom message back to our component.