compose box: Use standard Redux dispatch, cutting `connectWithActions`.
This is a pilot of a refactor that we might apply broadly across the
codebase.  Our `connectWithActions` helper is a small wrapper around
Redux's `bindActionCreators`, and we apply it to lots of our
components and then make a call like
  actions.addToOutbox(/* ... args to action ... */);
to cause an action to be dispatched.

This works, but it isn't the standard Redux pattern; it adds
additional layers of indirection that have to be worked through to
understand what's causing what to run when, layers that aren't found
in normal Redux examples or docs.  And in fact, upstream's doc for
`bindActionCreators` recommends limiting it to specific circumstances
that don't generally apply for us:
  https://redux.js.org/api-reference/bindactioncreators

So, take ComposeBox as an example component, and remove entirely the
use of `connectWithActions` around it.  Replace it with plain upstream
`connect` and the use of `dispatch` to dispatch actions.

This change turns out to have a few side effects, all of which I'm
pretty happy about:

* Call sites of `ComposeBoxContainer` now get properly type-checked!
  This is a very welcome bonus.  The `connectWithActions` helper wasn't
  correctly passing types through, while `connect` does.  Here, this
  surfaces a type error where different parts of our code disagree on
  whether the `narrow` property to these components should be optional;
  for this commit, just mark that error to be ignored explicitly.

* We have to specifically import the individual action creators in the
  module where we use them.  This is a bit more typing... but because it
  eliminates several indirections between the calling and called code,
  it makes it much simpler for both humans and machines to navigate to
  the action creator's definition.  In particular, in VS Code the
  "Go To Definition" (F12) feature now works, which it didn't before.

* Similarly, the hover info in VS Code now shows me the specific types
  that e.g. `addToOutbox` takes as parameters, rather than the alias
  `AddToOutboxActionCreator`, which is opaque to me and I have to go
  look up what it means by hand.