eslint: Fix how we break lines at operators.
This is something Prettier gets wrong, but thanks to prettier-eslint
we can now fix.

prettier-eslint internally consists of two steps: prettier, and
`eslint --fix`.  When the latter fixes a case of this, it can make the
later line a few characters longer.  That causes me to notice that
we're trying to have ESLint enforce a max line length, because we have
one spot where it causes a lint error.  So also turn that off, in
favor of letting prettier do prettier's job.

This change is entirely mechanical outside of the ESLint config;
just reran `yarn run prettier`, which runs prettier-eslint.