types: Convert those new doc comments to proper JSDoc.
We don't currently have much that automatically processes these, but
it seems to be the single common standard for documentation comments
in the JavaScript world, which any tool we might use will expect.
And it's super easy to convert these comments to it, so sure why not.

The main thing that this does right now, in the absence of tools, is
signal that these particular comments are *documentation comments*:
that they're meant as prose explaining the interface to the reader,
rather than explaining implementation details or providing
miscellaneous annotations like $FlowFixMe.

There's at least one other thing it does: for me in VS Code, these
doc comments now sometimes (though not always) show up as a helpful
annotation when I hover on a reference to the type elsewhere.