eslint: Remove case exception to `import/no-unresolved`.
The default configuration for this is
  [error, { caseSensitive: true, commonjs: true }]

The `commonjs` option only turns on some additional checks that would
trigger if we used commonjs imports; it has no effect for us.

The `caseSensitive` option turns on a check when the local filesystem
is case-insensitive (like the default macOS and Windows filesystems),
which makes sure that the filename is actually correct even with its
case.  This seems like an appropriate check; follow the default by
leaving it on.