realm input: Allow second-level domains like `example.org`.
Previously we expected that the Zulip server URL will always be in the
format of `subdomain.domain.tld`.  This is a mistake as some users are
hosting it on `domain.tld`.

Now the helper function `autocompleteUrl` no longer has the
`shortAppend` parameter and never tries to expand a string in the
form of `something.something`.