prettier: Replace prettier-eslint with just prettier then eslint.
This is oddly much faster: on my laptop, running on our whole
codebase, 10s + 5s = 15s, where `prettier-eslint` takes 36s.
I don't know where the other 21s in `prettier-eslint` are spent,
but this makes that not really matter to us.

We keep `prettier-eslint` in the `test:prettier` case, because it's
a bit more convenient than doing the check ourselves and because
that case is only relevant in CI, where an extra 21s matters less.