edit stream: Split create and edit logic
In order to add more different behavior to the create stream flow,
instead of overloading the behavior of `EditStreamCard` with more
'if'-s, keep just the common code there.

Now `EditStreamCard` is a general-purpose UI that on successful
submission calls `onComplete`.

`CreateStreamScreen` and `EditSreamScreen` now retrieve the relevant
data from the store and send the Redux actions on success.

[greg: Removed unused `streamId` prop on CreateStreamScreen.]