webview: Use inline-block to display emoji reactions.
This eliminates some excess padding that appeared on the right of each
reaction other than the last on a line.  Details:
  https://github.com/zulip/zulip-mobile/pull/2656#discussion_r194213679

With the different layout model, a large `line-height` no longer has
the desired effect, so we replace it with vertical margin and more
padding.

Also remove the `line-height` on the parent element, which has no
effect, and add a bit more horizontal padding still.