stream settings: Add InviteUsersScreen to add subscribers.
After creating a new stream on mobile, the user previously was
not able to invite people to it, thus making this feature useless.

This is a new screen that shows the familiar user list from the
'create group' UI that allows selecting people (now extracted
into a `UserPickerCard` component)

On a successful selection we use `subscriptionAdd` to add the
selected people to the stream.