settings: Fix navigation after actions/dispatch refactor.
In a4b9789c2 "settings card: Use standard Redux dispatch.", a function
call was mistakenly dropped -- we wrapped each of these expressions in
a lambda without inserting the corresponding call to correct for it.
The effect was that these buttons just didn't do anything, so it
wasn't possible to reach the inner settings screens.