flow: Add test suite for types.
This allows us to verify that our Flow types work as intended.
In particular, we can now assert type declarations that should
error.

Without additional configuration, Flow prints assertion errors
as warnings.  Add `--max-warnings 0` to make all such warnings
errors, which makes CI tests fail when our tests fail.

Also, add a test for the `Action` type.

Disable eslint for these tests, because we are not testing the
code but the types, and unused variables and the like are
acceptable for this.