docs: Write a few words about our own RN repo.
Because zulip/react-native doesn't have a single particularly
canonical master branch -- instead, it'll be a collection of short
branches on top of a succession of different upstream releases -- it
isn't a natural place to try to maintain much documentation.  Instead,
put an explanation here, which we can of course expand over time.

Over in zulip/react-native, we can then just have a short note in the
README pointing to this.