persist: Cache `streams` and `userGroups`.
This makes the app more usable when offline.

Also, move `users` from `storeKeys` to `cacheKeys`.  This does not
change any behavior, as keys in either of these lists are treated the
same; but it makes it more 'conceptually' correct, as we are caching
users and not storing them locally as a single source of truth.