notif android: Don't be too specific about stream notifs.
When we're notified on a stream message, it might be an @-mention --
or it might not, as the user can choose to get notifications on all
messages in a given stream.

See screenshot in #2666 and discussion there.  This is a minimal fix
that keeps the quality pretty high when it really is an @-mention.

Ideally we probably want to be taking the text here from the server,
e.g. from the `alert` field on the notification bundle, but the
quality of the text we currently put there isn't great, and in
particular when the message really is an @-mention it'd represent a
substantial regression in the UX.  Hence this fix instead, for now.