refactor: Merge Touchable.android and Touchable.ios
While splitting code in two files depending on the platfrom is
generally a good idea, in the case of Touchable there is too
little unique code and too much duplication.

[greg: I think this is actually an improvement already, even
 before the additional code added later!  Discussion in #2753.]

In anticipation of adding some more code that is exactly the same
it is apparent merging them and `if`-ing the differences is better.

While some improvements can be made, this refactor keeps the behavior
strictly the same but simplifies the logic and makes it more obvious.

Instead of WrapperComponent we just check if the component has
any onPress or onLongPress event attached, if not returns a `View`