api: Report errors to caller in a more useful way.
We have two separate error paths in processing an API response, with
rather confusingly similar jobs.  Their behavior works out to be that
if the server gives us a response with a non-2xx HTTP status code,
then we skip trying to parse the body as JSON.

The trouble with this behavior is that it's sometimes important to
know more information than "there was an error".  The Zulip API
handles this nicely by always returning JSON in the response, and
including useful information there.  So we need to parse that JSON.
Do that, and include the result in a thrown error just like we do in
returning a result.

Also take the opportunity to rearrange the interface for how our `api`
code provides error information to the caller.  For application-level
code, the entire JSON response is potentially relevant, just like it
is for non-error responses, so we return the whole thing.  On the
other hand, the Response object from the Fetch API is inappropriately
full of system-level details, so we take it out.

There was just one call site, in `startEventPolling`, where we
actually tried to inspect an error -- and the condition there would
actually never fire, regardless of what response we got.  This meant
that when our event queue on the server expired, we'd never notice and
get a new one, so the data in the app would turn stale: i.e., #2800.
So we make that site use the new API; it should work now.

Fixes #2800.