AnimatedComponent: Also animate on component creation.
Previously, AnimatedComponent.animate() was only called in the
componentDidUpdate() method, which is not called on the creation of
the component.  This works fine for our current uses of
AnimatedComponent, but it won't work for components that need to be
animated right after they have been created.  An example of this is
OfflineNotice, which will be animated in the next commit.

Currently, only ComponentMenu uses AnimatedComponent. It is not visibly
affected by this change, because it is hidden right after creation.