types: Specify that User.avatar_url can be null.
Based on empirical observation of the state. This does not require any
changes to our code apart from propagating the type correction and ensuring
that `getMediumAvatar` doesn't get called with a `null` avatar url. The
one place where a valid avatar url is actually needed is Avatar.js. Here,
if avatar_url is null, a gravatar is generated instead.